+++ changelog/changelog.md
@@ -13,3 +13,9 @@
 ## 20/12/22 Changelog
 - Change user class admin variable to accountType to handle 3 different user types (Customer, doctor and delivery)
 - formHelper now supports wtforms RadioField
+
+## 21/12/22 Changelog
+- Done admin change password page and functionality
+- Done zoom functionality
+- Created delivery address class
+- Removed delivery info fields from edit/create user pages